Alternative approach to DeferredWorkManager internal logic
